+++ R/assert.r
@@ -14,10 +14,8 @@
 #' @param combine [\code{character(1)}]\cr
 #'  \dQuote{or} or \dQuote{and} to combine the check functions with an OR
 #'  or AND, respectively.
-#' @param .var.name [character(1)]\cr
-#'  Name of object to check. Defaults to a heuristic to determine
-#'  the name of the first argument of the first call.
-#' @return Throws an error if all checks fails and invisibly returns
+#' @template var.name
+#' @return Throws an error if all checks fail and invisibly returns
 #'  \code{TRUE} otherwise.
 #' @export
 #' @examples
@@ -27,7 +25,7 @@
 #' x = 1
 #' assert(checkChoice(x, c("a", "b")), checkDataFrame(x))
 #' }
-assert = function(..., combine = "or", .var.name) {
+assert = function(..., combine = "or", .var.name = NULL) {
   assertChoice(combine, c("or", "and"))
   dots = match.call(expand.dots = FALSE)$...
   env = parent.frame()
@@ -35,11 +33,11 @@ assert = function(..., combine = "or", .var.name) {
     msgs = character(length(dots))
     for (i in seq_along(dots)) {
       val = eval(dots[[i]], envir = env)
-      if (isTRUE(val))
+      if (identical(val, TRUE))
         return(invisible(TRUE))
       msgs[i] = as.character(val)
     }
-    if (missing(.var.name))
+    if (is.null(.var.name))
       .var.name = as.character(dots[[1L]])[2L]
     if (length(msgs) > 1L)
       msgs = sprintf("%s: %s", lapply(dots, function(x) as.character(x)[1L]), msgs)
@@ -47,8 +45,8 @@ assert = function(..., combine = "or", .var.name) {
   } else {
     for (i in seq_along(dots)) {
       val = eval(dots[[i]], envir = env)
-      if (!isTRUE(val)) {
-        if (missing(.var.name))
+      if (!identical(val, TRUE)) {
+        if (is.null(.var.name))
           .var.name = as.character(dots[[1L]])[2L]
         mstop(qamsg(NULL, val, .var.name, recursive = FALSE))
       }